feat(condo): DOMA-9367 rename to ServiceProblemsAlert
