+++ apps/condo/domains/common/components/TopNotifications.tsx
@@ -8,7 +8,7 @@ import { Button } from '@condo/domains/common/components/Button'
 import { colors } from '@condo/domains/common/constants/style'
 
 import { useLayoutContext } from './LayoutContext'
-import { NotWorkingAlert } from './NotWorkingAlert'
+import { ServiceProblemsAlert } from './ServiceProblemsAlert'
 
 
 const notificationAlert = ({ isSmall }) => css`
@@ -52,7 +52,7 @@ interface ITopNotificationHookResult {
     addNotification: (notification: ITopNotification) => void,
 }
 
-export const useTopNotificationsHook = (withNotWorkingNotification?: boolean): ITopNotificationHookResult => {
+export const useTopNotificationsHook = (withServiceProblemsAlert?: boolean): ITopNotificationHookResult => {
     const [topNotifications, setTopNotifications] = useState<ITopNotification[]>([])
     const addNotification = (notification: ITopNotification) => {
         if (!topNotifications.find(existedNotification => existedNotification.id === notification.id)) {
@@ -66,12 +66,12 @@ export const useTopNotificationsHook = (withNotWorkingNotification?: boolean): I
     const TopNotificationComponent: React.FC = () => {
         const { breakpoints } = useLayoutContext()
 
-        if (topNotifications.length === 0 && !withNotWorkingNotification) return null
+        if (topNotifications.length === 0 && !withServiceProblemsAlert) return null
 
         return (
             <>
                 <Affix>
-                    {withNotWorkingNotification && <NotWorkingAlert />}
+                    {withServiceProblemsAlert && <ServiceProblemsAlert />}
                     {
                         topNotifications.map(notification => {
                             return (
