+++ apps/condo/pages/_app.tsx
@@ -466,7 +466,7 @@ const MyApp = ({ Component, pageProps }) => {
                     <SetupTelegramNotificationsBanner />
                     <GlobalStyle/>
                     {shouldDisplayCookieAgreement && <CookieAgreement/>}
-                    <LayoutContextProvider withNotWorkingNotification>
+                    <LayoutContextProvider withServiceProblemsAlert>
                         <TasksProvider>
                             <PostMessageProvider>
                                 <TrackingProvider>
